Refactor publish workflow helpers (#230)
* Align publish helpers with review feedback

* Refactor publish workflows helper

* Modularise publish check tests

* Introduce crate processing config

* Refactor live publish integration test helpers

* Refactor dry-run orchestration test helpers

* Refine integration test setup helpers
