+++ Makefile
@@ -20,7 +20,7 @@ test: ## Run tests with warnings treated as errors
 	RUSTFLAGS="-D warnings" $(CARGO) test --workspace --all-targets --all-features $(BUILD_JOBS)
 	# Exercise the Python release automation alongside the Rust suite.
 	$(UV) run --with pytest --with cyclopts --with plumbum --with tomlkit \
-		python -m pytest scripts/tests/test_run_publish_check.py
+		python -m pytest scripts/tests/publish_check
 
 target/%/$(APP): ## Build binary in debug or release mode
 	$(CARGO) build $(BUILD_JOBS) $(if $(findstring release,$(@)),--release) --bin $(APP)
