+++ scripts/tests/publish_check/test_live_publish_workflow.py
@@ -1,0 +1,95 @@
+"""Live publish workflow orchestration tests."""
+
+from __future__ import annotations
+
+from pathlib import Path
+from types import ModuleType
+
+import pytest
+
+
+def test_process_crates_for_live_publish_delegates_configuration(
+    monkeypatch: pytest.MonkeyPatch,
+    tmp_path: Path,
+    run_publish_check_module: ModuleType,
+) -> None:
+    captured: dict[str, object] = {}
+
+    def fake_process_crates(
+        workspace: Path,
+        timeout_secs: int,
+        config: object,
+        crate_action: object,
+    ) -> None:
+        captured["workspace"] = workspace
+        captured["timeout"] = timeout_secs
+        captured["config"] = config
+        captured["crate_action"] = crate_action
+
+    monkeypatch.setattr(run_publish_check_module, "_process_crates", fake_process_crates)
+
+    workspace = tmp_path / "live"
+    run_publish_check_module._process_crates_for_live_publish(workspace, 99)
+
+    assert captured["workspace"] == workspace
+    assert captured["timeout"] == 99
+    config = captured["config"]
+    assert isinstance(config, run_publish_check_module.CrateProcessingConfig)
+    assert config.strip_patch is False
+    assert config.include_local_path is False
+    assert config.apply_per_crate is True
+    assert config.per_crate_cleanup is run_publish_check_module.remove_patch_entry
+    assert captured["crate_action"] is run_publish_check_module.publish_crate_commands
+
+
+def test_process_crates_for_live_publish_runs_publish_workflow(
+    monkeypatch: pytest.MonkeyPatch,
+    tmp_path: Path,
+    run_publish_check_module: ModuleType,
+) -> None:
+    workspace = tmp_path / "workspace"
+    workspace.mkdir()
+    (workspace / "Cargo.toml").write_text("", encoding="utf-8")
+
+    calls: list[tuple[str, object]] = []
+
+    monkeypatch.setattr(run_publish_check_module, "strip_patch_section", lambda *_: None)
+    monkeypatch.setattr(run_publish_check_module, "workspace_version", lambda _m: "0.1.0")
+
+    def fake_apply(
+        root: Path,
+        version: str,
+        *,
+        include_local_path: bool,
+        crates: tuple[str, ...] | None = None,
+    ) -> None:
+        calls.append(("apply", (root, version, include_local_path, crates)))
+
+    monkeypatch.setattr(run_publish_check_module, "apply_workspace_replacements", fake_apply)
+
+    def fake_publish(crate: str, root: Path, *, timeout_secs: int) -> None:
+        calls.append(("publish", (crate, root, timeout_secs)))
+
+    monkeypatch.setattr(run_publish_check_module, "publish_crate_commands", fake_publish)
+
+    def fake_remove(manifest: Path, crate: str) -> None:
+        calls.append(("remove_patch", (manifest, crate)))
+
+    monkeypatch.setattr(run_publish_check_module, "remove_patch_entry", fake_remove)
+    monkeypatch.setattr(
+        run_publish_check_module,
+        "CRATE_ORDER",
+        ("crate-a", "crate-b"),
+    )
+
+    run_publish_check_module._process_crates_for_live_publish(workspace, 42)
+
+    manifest = workspace / "Cargo.toml"
+    assert calls == [
+        ("apply", (workspace, "0.1.0", False, ("crate-a",))),
+        ("publish", ("crate-a", workspace, 42)),
+        ("remove_patch", (manifest, "crate-a")),
+        ("apply", (workspace, "0.1.0", False, ("crate-b",))),
+        ("publish", ("crate-b", workspace, 42)),
+        ("remove_patch", (manifest, "crate-b")),
+    ]
